+++ tptplus/CHANGELOG.md
@@ -23,4 +23,5 @@
 - v0.1.8 Fixed issue where some greater-than-signs are not handled correctly by command "Format TPTP File".
 - v0.1.9 Added local pretty-printer so that "Format TPTP File" works without Internet connection. The original remote pretty-printer is kept as a fallback.
 - v0.1.10 Added error handling for "Format TPTP File": display error message and jump to position of syntax error.
-- v0.1.11 Added title-menu button for "TPTP: Check Syntax and Format TPTP File"; improved error handling for pretty-printer.
+- v0.1.11 Added title-menu button for "TPTP: Pretty-Print TPTP File"; improved error handling for pretty-printer.
+- v0.1.12 Added title-menu button for "TPTP: Type-Check with LEO-III"; improved error handling for pretty-printer; retired legacy formatter.